added more checks
Needed to add some checks here. sqlite3 check would actually create an empty file if it didn't already exist and report it as "ok". To work around this we check if the file exists first AND then proceed to check it for errors. Also, added an email check for the root user so it can send an email, if not it will output the db errors to just the shell.

+++ db_check.zsh
@@ -2,22 +2,31 @@
 
 # Checks TrueNAS database consistency, and if there's an error, sends an email to the root user
 
-#The command we use to check database consistency
-DBCHECK=$(sqlite3 /data/freenas-v1.db 'pragma integrity_check')
+DB="/data/freenas-v1.db"
+DBCHECK=$(sqlite3 "$DB" 'pragma integrity_check')
+TMPFILE=$(mktemp)
+EMAIL=$(awk '{ if ($1 == "root\:") {print $2 } }' /etc/aliases)
+HOSTNAME=$(hostname)
 
-if [[ "$DBCHECK" == "ok" ]]; then
-    echo Database Check OK
+if [[ -f "$DB" && "$DBCHECK" == "ok" ]]; then
+    echo "Database Check OK"
   else
-    echo Database Check FAILED!
-    TMPFILE=$(mktemp)
-    EMAIL=$(awk '{ if ($1 == "root\:") {print $2 } }' /etc/aliases)
-    HOSTNAME=$(hostname)
-    {
-    printf "To: %s\n" "$EMAIL"
-    printf "Subject: ERROR: Database corrupt on server %s\n\n" "$HOSTNAME"
-    printf "%s has a corrupt TrueNAS config. Please restore from a backup ASAP!\n\n\nErrors:\n\n" "$HOSTNAME"
-    printf "%s" "$DBCHECK"
-    } >> "$TMPFILE"
-    sendmail -t < "$TMPFILE"
-    rm "$TMPFILE"
+    if [[ ! -f "$DB" ]]; then
+        echo "Database not present. Please check path."
+      else
+        if [[ "$EMAIL" == "None" ]]; then
+            echo "Database error was found, but no email is configured for root user"
+            echo "Outputting error to shell only"
+            echo "ERROR: '$DBCHECK'"
+          else
+            {
+            printf "To: %s\n" "$EMAIL"
+            printf "Subject: ERROR: Database corrupt on server %s\n\n" "$HOSTNAME"
+            printf "%s has a corrupt TrueNAS config. Please restore from a backup ASAP!\n\n\nErrors:\n\n" "$HOSTNAME"
+            printf "%s" "$DBCHECK"
+            } >> "$TMPFILE"
+            sendmail -t < "$TMPFILE"
+            rm "$TMPFILE"
+        fi
+    fi
 fi
